Move NewsViewAdapter to separate file
