+++ TopNews/TopNewsiOS/NewsUI/Composers/NewsUIComposer.swift
@@ -31,26 +31,3 @@ private extension NewsViewController {
         return newsController
     }
 }
-
-private final class NewsViewAdapter: NewsView {
-    private weak var controller: NewsViewController?
-    private let imageLoader: NewsImageDataLoader
-    
-    init(controller: NewsViewController, imageLoader: NewsImageDataLoader) {
-        self.controller = controller
-        self.imageLoader = imageLoader
-    }
-    
-    func display(_ viewModel: NewsViewModel) {
-        controller?.tableModel = viewModel.news.map { model in
-            let adapter = NewsImageDataLoaderPresentationAdapter<WeakRefVirtualProxy<NewsImageCellController>, UIImage>(model: model, imageLoader: imageLoader)
-            let view = NewsImageCellController(delegate: adapter)
-            
-            adapter.presenter = NewsImagePresenter(
-                view: WeakRefVirtualProxy(view),
-                imageTransformer: UIImage.init)
-            
-            return view
-        }
-    }
-}
